added student view for booking meetings
