+++ ohs_gui/src/components/officeHours/OHContainer.js
@@ -6,13 +6,14 @@ import OHSlot from './OHSlot';
 
 class OHContainer extends PureComponent {
   render() {
-    const { bookedSlots, toggleBooking } = this.props;
+    const { bookedSlots, toggleBooking, showBooked } = this.props;
     const slots = bookedSlots.map((d, i) => (
       <OHSlot
         key={shortid.generate()}
         id={i}
         toggleBooking={toggleBooking}
         booked={bookedSlots[i]}
+        showBooked={showBooked}
       />
     ));
