+++ ohs_gui/src/components/officeHours/OHSlot.css
@@ -27,6 +27,7 @@
 }
 
 .booked {
+  color: gray;
   background-color: navy;
 }
